feat: add additional user information

+++ src/auth/auth.service.ts
@@ -8,11 +8,15 @@ import { LoginDTO } from './dto/login.dto';
 import { compare } from 'bcrypt';
 import { sign } from 'jsonwebtoken';
 import { TOKEN_KEY } from './jwt.token';
+import { AdditionalUserInfoEntity } from './schema/additionalUserInfo.schema';
+import { UpdateUserDTO } from './dto/updateUser.dto';
 
 @Injectable()
 export class AuthService {
   constructor(
     @InjectModel(UserEntity.name) private userModel: Model<UserEntity>,
+    @InjectModel(AdditionalUserInfoEntity.name)
+    private additionalUserInfoModel: Model<AdditionalUserInfoEntity>,
   ) {}
 
   async getAll(): Promise<any> {
@@ -21,15 +25,21 @@ export class AuthService {
   }
 
   async deleteUser(loginDto: LoginDTO): Promise<boolean> {
-    const user = await this.userModel.findOne({ email: loginDto.email });
+    const user = await this.userModel
+      .findOne({ email: loginDto.email })
+      .select('+password');
     if (!user) {
       throw new HttpException(
-        ' Email is already exist ',
+        ' User not found',
         HttpStatus.UNPROCESSABLE_ENTITY,
       );
     }
-    await this.userModel.deleteOne({ email: loginDto.email });
-    return true;
+    const isPasswordCorrect = await compare(loginDto.password, user.password);
+    if (isPasswordCorrect) {
+      await this.userModel.deleteOne({ email: loginDto.email });
+      return true;
+    } else
+      throw new HttpException('Password Uncorrect', HttpStatus.UNAUTHORIZED);
   }
 
   async createUser(createUserDto: CreateUserDTO): Promise<UserEntity> {
@@ -42,6 +52,10 @@ export class AuthService {
     }
     const createdUser = new this.userModel(createUserDto);
     return createdUser.save();
+  }
+
+  async getAllUserInfo(): Promise<any> {
+    return this.additionalUserInfoModel.find().exec();
   }
 
   async loginUser(loginDto: LoginDTO): Promise<UserEntity> {
@@ -79,5 +93,35 @@ export class AuthService {
   async findByEmail(email: string): Promise<UserEntity> {
     const user = this.userModel.findOne({ email: email });
     return user;
+  }
+
+  async updateUserInfo(
+    user: UserEntity,
+    updateUserDto: UpdateUserDTO,
+  ): Promise<AdditionalUserInfoEntity> {
+    const _user = await this.additionalUserInfoModel.findOne({ user: user });
+    console.warn(_user);
+    if (_user) {
+      const a = await this.additionalUserInfoModel.updateOne(
+        { user: user },
+        {
+          $set: {
+            user: user,
+            city: updateUserDto.city,
+            icon: updateUserDto.icon,
+            login: updateUserDto.login,
+            surname: updateUserDto.surname,
+            role: updateUserDto.role,
+          },
+        },
+      );
+      console.log('----->', a);
+    } else {
+      const createdUser = new this.additionalUserInfoModel({
+        user: user,
+        ...updateUserDto,
+      });
+      return createdUser.save();
+    }
   }
 }
